debugged drop logic
