+++ waste-of-time/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/App.js":"1","/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/pages/Game-Container.js":"2","/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/index.js":"3","/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/reportWebVitals.js":"4"},{"size":176,"mtime":1610936602044,"results":"5","hashOfConfig":"6"},{"size":2354,"mtime":1610939262116,"results":"7","hashOfConfig":"6"},{"size":501,"mtime":1610938897423,"results":"8","hashOfConfig":"6"},{"size":362,"mtime":1610932411469,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"qdysyn",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"19"},"/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/App.js",[],["20","21"],"/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/pages/Game-Container.js",[],"/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/index.js",[],"/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/reportWebVitals.js",[],["22","23"],{"ruleId":"24","replacedBy":"25"},{"ruleId":"26","replacedBy":"27"},{"ruleId":"24","replacedBy":"28"},{"ruleId":"26","replacedBy":"29"},"no-native-reassign",["30"],"no-negated-in-lhs",["31"],["30"],["31"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/victoria/development/recycle-game-sandbox/springdrag/src/index.js":"1","/Users/victoria/development/sandbox/springdrag2/src/index.js":"2","/Users/victoria/development/sandbox/springdrag2/src/App.js":"3","/Users/victoria/development/sandbox/springdrag2/src/pages/Game.js":"4","/Users/victoria/development/sandbox/springdrag2/src/reportWebVitals.js":"5","/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/index.js":"6","/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/reportWebVitals.js":"7","/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/App.js":"8","/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/pages/Game.js":"9"},{"size":2074,"mtime":1610944406116,"results":"10","hashOfConfig":"11"},{"size":875,"mtime":1610946246527,"results":"12","hashOfConfig":"13"},{"size":152,"mtime":1610946355916,"results":"14","hashOfConfig":"13"},{"size":1851,"mtime":1610946520535,"results":"15","hashOfConfig":"13"},{"size":362,"mtime":1610945269896,"results":"16","hashOfConfig":"13"},{"size":875,"mtime":1610946861972,"results":"17","hashOfConfig":"18"},{"size":362,"mtime":1610946861976,"results":"19","hashOfConfig":"18"},{"size":152,"mtime":1610946861926,"results":"20","hashOfConfig":"18"},{"size":1851,"mtime":1610946861975,"results":"21","hashOfConfig":"18"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"13f76wu",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"okttdj",{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"qdysyn",{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/victoria/development/recycle-game-sandbox/springdrag/src/index.js",[],"/Users/victoria/development/sandbox/springdrag2/src/index.js",[],"/Users/victoria/development/sandbox/springdrag2/src/App.js",[],"/Users/victoria/development/sandbox/springdrag2/src/pages/Game.js",[],"/Users/victoria/development/sandbox/springdrag2/src/reportWebVitals.js",[],"/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/index.js",[],"/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/reportWebVitals.js",[],"/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/App.js",[],"/Users/victoria/development/recycle-game-sandbox/waste-of-time/src/pages/Game.js",[]]
