+++ waste-of-time/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "waste-of-time",
+  "name": "springdrag",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -9,7 +9,7 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.1",
-    "react-spring": "^8.0.27",
+    "react-spring": "9.0.0-rc.2",
     "react-svg": "^11.2.2",
     "react-use-gesture": "^9.0.0",
     "web-vitals": "^0.2.4"
