+++ waste-of-time/src/App.css
@@ -1,38 +1,68 @@
-.App {
-  text-align: center;
+body,
+html {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, avenir next, avenir;
+  background: #f0f0f0;
+  height: 100vh;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+* {
+  box-sizing: border-box;
+  user-select: none;
+  -moz-user-select: none;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+#root {
   display: flex;
-  flex-direction: column;
   align-items: center;
   justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+  height: 100vh;
+  width: 100vw;
 }
 
-.App-link {
-  color: #61dafb;
+#root > div {
+  /* position: relative;
+  width: 30vw;
+  height: 30vw;
+  background: grey;
+  border-radius: 5px;
+  box-shadow: 0px 10px 30px -5px rgba(0, 0, 0, 0.3); */
+  transition: box-shadow 0.5s, opacity 0.5s;
+  will-change: transform;
+  /* border: 10px solid white; */
+  cursor: grab;
+  overflow: hidden;
+  touch-action: none;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+#root > div > div {
+  will-change: transform;
+  height: 100%;
+  margin: 0vw 0;
+}
+
+#root > div > div > * {
+  height: 100%;
+  background-size: cover;
+  background-position: center center;
+  margin: 0vw 0;
+}
+
+#root > div.dragging {
+  cursor: grabbing;
+}
+
+#root > div.disabled {
+  opacity: 0.5;
+  cursor: default;
+}
+
+/* #root > div:hover {
+  box-shadow: 0px 30px 100px -10px rgba(0, 0, 0, 0.4);
+} */
+
+button {
+  position: absolute;
+  top: 20px;
+  left: 20px;
 }
