+++ waste-of-time/src/index.js
@@ -1,17 +1,33 @@
+// import React, { useRef, useEffect, useState } from "react";
+// import ReactDOM from "react-dom";
+// import { useSpring, animated } from "react-spring";
+// import { useGesture } from "react-use-gesture";
+// import { ReactSVG } from "react-svg";
+// import "./App.css";
+
+// import Bone from "./assets/fish_bone.svg"
+// import Bin from "./assets/bin_closed.svg"
+
+
+// const calcX = (y, ly) => -(y - ly - window.innerHeight / 2) / 20;
+// const calcY = (x, lx) => (x - lx - window.innerWidth / 2) / 20;
+
+// document.addEventListener("gesturestart", (e) => e.preventDefault());
+// document.addEventListener("gesturechange", (e) => e.preventDefault());
+
+
 import React from 'react';
 import ReactDOM from 'react-dom';
-// import './index.css';
-import App from './App';
-// import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './reportWebVitals';
+import App from "./App"
 
-const rootElement = document.getElementById('root')
+
 
 ReactDOM.render(
-  
-    <App />, rootElement)
-  
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();
+  <App />
+  ,
+  document.getElementById('root')
+);
+
+reportWebVitals();
